+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### 🚀 Added
+### ✏️ Changed
 
-- New Feature: Random groups
-- New Feature: Questionnaires can be only visible for specific groups
+- Improved app_usage item handling.
+- Improved access key handling in study list.